Track specific mobile/object created by a zone reset rule.
Load limit numbers don't work well when zones have multiples of the same mobile/object.  By remember
which rule created a certain mobile or object, it is possible to avoid using a rule where the object or
mobile already exists.  This prevents 2 of the same mobile loading up in one location when only one should.